Add normalized fuzzy matchers for known types (string, EnPronunciation, DistanceInput)
Add normalized fuzzy matchers for known types. Specialization is not defined in C# because generics are resolved on runtime. Hence, I used derived class instead. This implementation leads the way to pass the distance operators to native functions (instead of passing managed callbacks).

Another approach would be to define Extensions for FindNearest* methods but that implies to always pass the phoneticWeightPercentage as a parameter for the DistanceInput case or to expose the field in an undesired way in the interface. i.e.:
        public static IList<Match<Target>> FindNearestWithinNormalized<Target>(this IFuzzyMatcher<Target, DistanceInput> matcher, DistanceInput query, double limit, int count, double phoneticWeightPercentage)